requirements.txt
asgiref==3.5.0
autopep8==1.7.0
azure-core==1.26.0
backoff==2.2.1
beautifulsoup4==4.11.1
bootstrap4==0.1.0
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
comtypes==1.1.11
cryptography==38.0.1
decorator==5.1.1
distlib==0.3.6
Django==4.0.2
django-bootstrap4==22.2
django-cleanup==7.0.0
django-crispy-forms==1.14.0
django-filter==22.1
django-storages==1.13.1
django-utils2==2.12.1
django-weekday-field==1.1.0
dlib==19.24.0
docopt==0.6.2
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.8.0
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
geopy==2.2.0
idna==3.3
isodate==0.6.1
mock==4.0.3
msrest==0.7.1
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.29
mysqlclient==2.1.1
networkx==2.8.7
nominatim==0.1
numexpr==2.8.3
numpy==1.22.4
oauthlib==3.2.1
opencage==2.0.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.3
Pillow==9.1.1
pipenv==2022.10.10
pipreqs==0.4.11
platformdirs==2.5.2
playsound==1.3.0
psycopg2==2.9.4
pybbn==3.2.1
pycodestyle==2.9.1
pycparser==2.21
pyodbc==4.0.32
pyOpenSSL==22.1.0
pyparsing==3.0.9
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.8.2
python-utils==3.3.3
pyttsx3==2.90
pytz==2022.1
pywin32==304
ratelim==0.1.6
requests==2.28.1
requests-oauthlib==1.3.1
scipy==1.9.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tables==3.7.0
tk==0.1.0
toml==0.10.2
typing_extensions==4.4.0
tzdata==2021.5
urllib3==1.26.12
virtualenv==20.16.5
virtualenv-clone==0.5.7
win-inet-pton==1.1.0
wincertstore==0.2
xlwt==1.3.0
yarg==0.1.9
